remove withMutations from cookies reducer, move utils/cookies to logic/cookies
